Move modified ref test server to own file (#2725)
Moved the modified reference server from the unit test to its own file. 
Also moved the modified server side session, the modified session manager and the modified master node manager to the same file.
